requirements.txt
colorgram.py==1.2.0
Flask==2.0.1
flask-restx==0.5.1
urllib3==1.26.6
webcolors==1.3
Werkzeug==2.0.1